requirements.txt
streamlit
pillow>=10.0.0
opencv-python-headless
easyocr
setuptools>=68.0.0
wheel
pip>=23.0
fastapi
uvicorn[standard]
python-telegram-bot==20.3
python-dotenv
httpx
pandas>=1.4.0
numpy>=1.21.0
soundfile>=0.11.0
webrtcvad>=2.0.10
av>=10.0.0
streamlit-webrtc>=0.60.0
openai-whisper>=1.1.10
